refactor 각스더 폴더 이동
각스더 관련 폴더 이동
구조 변경 준비
