+++ src/page/gaksder/result/resultHead.js
